requirements.txt
azure-common==1.1.28
azure-storage-blob==12.18.3
bleach==6.1.0
boto3==1.28.63
celery==5.3.4
Django==4.2.6
django-celery-beat==2.5.0
django-debug-toolbar==4.2.0
django-solo==2.1.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-webpack-loader==2.0.1
gunicorn==21.2.0
numpy==1.26.0
openpyxl==3.1.2
pandas==2.1.1
Pillow==10.0.1
psycopg2==2.9.9
Pygments==2.16.1
python-magic==0.4.27
redis==5.0.1
requests==2.31.0
sentry-sdk==1.32.0
tqdm==4.66.1
ua-parser==0.18.0
